LayoutTests/accessibility/mac/text-marker-range-for-node-without-children-expected.txt

a
b
This tests that creating TextMarkerRange from nodes that have no children won't collapse to one side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textbox.textMarkerRangeLength(textboxRange) == 1 is true
PASS textbox.accessibilityElementForTextMarker(startMarker).description is 'AXDescription: parent level'
PASS textbox.accessibilityElementForTextMarker(endMarker).description is 'AXDescription: parent level'
PASS successfullyParsed is true

TEST COMPLETE
